user_center/requirements.txt
alembic
Flask
Flask-Cors
Flask-RESTful
PyMySQL
SQLAlchemy
requests
pika
redis
bcrypt
certifi
pyOpenSSL
Mako
python-redis-lock
aliyun-python-sdk-core-v3
aliyun-python-sdk-dysmsapi
pillow
PyYAML
oss2
boto3
awscli